django_backend/requirements.txt
Django>=3.1.7
djangorestframework
django-cors-headers
django-environ
psycopg2-binary
